Added dotenv to hide important variables and login info

+++ Backend/controllers/sauce.js
@@ -49,10 +49,12 @@ exports.modifySauce = async (req, res) => {
         const foundSauce = await Sauce.findById(id);
         if (foundSauce.userId !== req.auth.userId) res.status(403).json({ message: "403: unauthorized request." })
         else {
-            fs.unlink(`images/${foundSauce.imageUrl.split('/images/')[1]}`, (err) => {
-                if (err) throw err
-                console.log('image deleted successfully.');
-            })
+            if (req.file) {
+                fs.unlink(`images/${foundSauce.imageUrl.split('/images/')[1]}`, (err) => {
+                    if (err) throw err
+                    console.log('image deleted successfully.');
+                })
+            }
             await Sauce.findByIdAndUpdate(id, updatedSauce);
             res.status(200).json({ message: "Sauce modifiée !" })
         }
